refactor: Modularizado o arquivo events.js, junto da correção da apresentação dos eventos
